requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
Babel==2.10.3
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
cloudinary
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dj-database-url==1.0.0
Django==4.0.6
django-autoslug==1.9.8
django-channels==0.7.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-model-utils==4.2.0
django-notifications-hq 
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonfield==3.1.0
msgpack==1.0.4
oauthlib==3.2.0
phonenumbers==8.12.52
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python2==1.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
zope.interface==5.4.0
goslate==1.5.1
gunicorn==20.0.4
beautifulsoup4==4.9.3
bs4==0.0.1
chardet==3.0.4
click==7.1.2
django-rename-app==0.1.2
futures==3.1.1
goslate==1.5.1
gunicorn==20.0.4
hashids==1.3.1
lxml==4.6.2
PyDictionary==2.0.1
shortuuid==1.0.1
soupsieve==2.1